move server creation to constructor
