+++ src/ru/yandex/practicum/task/http/handlers/SubtasksHandler.java
@@ -1,30 +1,30 @@
-package ru.yandex.practicum.task.http.handles;
+package ru.yandex.practicum.task.http.handlers;
 
 import com.google.gson.JsonSyntaxException;
 import com.sun.net.httpserver.HttpExchange;
 import ru.yandex.practicum.task.error.NotFoundException;
 import ru.yandex.practicum.task.error.TimeIntersectedException;
 import ru.yandex.practicum.task.interfaces.TaskManager;
-import ru.yandex.practicum.task.tasks.Task;
+import ru.yandex.practicum.task.tasks.Subtask;
 
 import java.util.Optional;
 
-public class TasksHandler extends BaseHttpHandler {
+public class SubtasksHandler extends BaseHttpHandler {
 
-    public TasksHandler(TaskManager taskManager) {
-        super(taskManager, "tasks");
+    public SubtasksHandler(TaskManager taskManager) {
+        super(taskManager, "subtasks");
     }
 
     @Override
     protected void handleGetItems(HttpExchange exchange) {
-        sendResponse(exchange, taskManager.getAllTasks());
+        sendResponse(exchange, taskManager.getAllSubtasks());
     }
 
     @Override
     protected void handleGetItemById(HttpExchange exchange) {
         try {
-            Task task = taskManager.getTask(getId(exchange));
-            sendResponse(exchange, task);
+            Subtask subtask = taskManager.getSubtask(getId(exchange));
+            sendResponse(exchange, subtask);
         } catch (NotFoundException e) {
             sendNotFound(exchange, e.getMessage());
         }
@@ -37,12 +37,12 @@ public class TasksHandler extends BaseHttpHandler {
             if (bodyStr.get().isBlank()) {
                 throw new JsonSyntaxException("Тело запроса не может быть пустым");
             }
-            Task task = gson.fromJson(bodyStr.get(), Task.class);
+            Subtask subtask = gson.fromJson(bodyStr.get(), Subtask.class);
             try {
-                if (task.getId() == null) {
-                    taskManager.createTask(task);
+                if (subtask.getId() == null) {
+                    taskManager.createSubtask(subtask);
                 } else {
-                    taskManager.updateTask(task);
+                    taskManager.updateSubtask(subtask);
                 }
                 sendSuccess(exchange);
             } catch (TimeIntersectedException e) {
@@ -59,7 +59,7 @@ public class TasksHandler extends BaseHttpHandler {
     protected void handleDeleteItem(HttpExchange exchange) {
         try {
             int id = getId(exchange);
-            taskManager.deleteTask(id);
+            taskManager.deleteSubtask(id);
             sendSuccess(exchange);
         } catch (NotFoundException e) {
             sendNotFound(exchange, e.getMessage());
