+++ src/ru/yandex/practicum/task/http/handlers/TasksHandler.java
@@ -1,29 +1,30 @@
-package ru.yandex.practicum.task.http.handles;
+package ru.yandex.practicum.task.http.handlers;
 
 import com.google.gson.JsonSyntaxException;
 import com.sun.net.httpserver.HttpExchange;
 import ru.yandex.practicum.task.error.NotFoundException;
+import ru.yandex.practicum.task.error.TimeIntersectedException;
 import ru.yandex.practicum.task.interfaces.TaskManager;
-import ru.yandex.practicum.task.tasks.Epic;
+import ru.yandex.practicum.task.tasks.Task;
 
 import java.util.Optional;
 
-public class EpicsHandler extends BaseHttpHandler {
+public class TasksHandler extends BaseHttpHandler {
 
-    public EpicsHandler(TaskManager taskManager) {
-        super(taskManager, "epics");
+    public TasksHandler(TaskManager taskManager) {
+        super(taskManager, "tasks");
     }
 
     @Override
     protected void handleGetItems(HttpExchange exchange) {
-        sendResponse(exchange, taskManager.getAllEpics());
+        sendResponse(exchange, taskManager.getAllTasks());
     }
 
     @Override
     protected void handleGetItemById(HttpExchange exchange) {
         try {
-            Epic epic = taskManager.getEpic(getId(exchange));
-            sendResponse(exchange, epic);
+            Task task = taskManager.getTask(getId(exchange));
+            sendResponse(exchange, task);
         } catch (NotFoundException e) {
             sendNotFound(exchange, e.getMessage());
         }
@@ -36,13 +37,17 @@ public class EpicsHandler extends BaseHttpHandler {
             if (bodyStr.get().isBlank()) {
                 throw new JsonSyntaxException("Тело запроса не может быть пустым");
             }
-            Epic epic = gson.fromJson(bodyStr.get(), Epic.class);
-            if (epic.getId() == null) {
-                taskManager.createEpic(epic);
-            } else {
-                taskManager.updateEpic(epic);
+            Task task = gson.fromJson(bodyStr.get(), Task.class);
+            try {
+                if (task.getId() == null) {
+                    taskManager.createTask(task);
+                } else {
+                    taskManager.updateTask(task);
+                }
+                sendSuccess(exchange);
+            } catch (TimeIntersectedException e) {
+                sendHasInteractions(exchange, e.getMessage());
             }
-            sendSuccess(exchange);
         } catch (NullPointerException | JsonSyntaxException e) {
             sendBadRequest(exchange, "Неверное тело запроса");
         } catch (Exception e) {
@@ -51,20 +56,10 @@ public class EpicsHandler extends BaseHttpHandler {
     }
 
     @Override
-    protected void handleGetChildrenItemsById(HttpExchange exchange) {
-        try {
-            Epic epic = taskManager.getEpic(getId(exchange));
-            sendResponse(exchange, taskManager.getSubtasksByEpic(epic));
-        } catch (NotFoundException e) {
-            sendNotFound(exchange, e.getMessage());
-        }
-    }
-
-    @Override
     protected void handleDeleteItem(HttpExchange exchange) {
         try {
             int id = getId(exchange);
-            taskManager.deleteEpic(id);
+            taskManager.deleteTask(id);
             sendSuccess(exchange);
         } catch (NotFoundException e) {
             sendNotFound(exchange, e.getMessage());
